Extract randomize and types
