+++ packages/scrabble-solver/src/components/PlainTiles/lib/randomize.ts
@@ -1,0 +1,3 @@
+const randomize = (value: number, maxChange: number): number => value + maxChange * 2 * (0.5 - Math.random());
+
+export default randomize;
